feat(home): themed inline Open Repository card for local backends
Replace the popup 'Open a Repository' dialog with an inline, themed card
that matches the cloud 'Open Repository' card (UI-kit Card/CardTitle +
BrandButton), so the local/self-hosted home screen looks consistent.

The card takes owner/repo or a clone URL plus an optional branch, clones
into the managed workspace dir via the runtime, and feeds the existing
local workspace -> conversation flow. Removes the now-unused
open-local-repository-dialog.tsx. Behaviour and clone logic unchanged;
all home-chat-launcher + parser unit tests still pass.
